Did some additional work and restructuring. Chiefly, sets up more to integrate with JD's use case, and edits the InputDialog to support passing additional data relating to the setup of the OverlapMaximization Use Case. Sets OverlapMaximizationController and TimeTableMatchInteractor up such that the other parts of the use case can listen for data from the InputDialog on things like the selected timetable.
Essentially, the skeleton is all done and I should hopefully just need to plug in JD's stuff now :) for intermediate input. And maybe link it to Kai's and add unit tests or something.

-- Kind of another commit, forgot to push that yesterday.

Also moved Flow.Subscriber() to the OverlapMaximizationController from team feedback. Also refined and completed the implementation of that a little bit more. Refactored TimeTableMatchInteractor accordingly so that it doesn't have to listen to the Dialog (a teammate told me that was a bad idea). Instead the Controller will listen to the Dialog, and then use that to call the Interactor's methods.

Interactor no longer stores mainTable, and timetables - instead, those are method parameters, because it doesn't make sense to make a new interactor / update existing one's fields for every new pair. This way we can make it a singleton!